// src/types/business.ts
export type businessDetails = {
    businessName: string;
    businessTitle: string;
    businessType: string;
    businessAddress: string;
    businessCity: string;
    businessState: string;
    businessPostalCode: number;
    businessPhone: number[];
    businessEmail: string;
    businessGSTIN: string;
};